changed the purpose of personal_diagnostic to be a barebones functional wrapper that just imports the personal module and returns the cfg dictionary for user purposes; moved the file getter to the example personal diag
